Added Argument command reporter without changing output
