+++ src/robotframework_find_unused/commands/arguments/__init__.py
@@ -1,0 +1,11 @@
+"""
+Find unused variables command
+"""
+
+from .arguments import cli_arguments
+from .options import ArgumentsOptions
+
+__all__ = [
+    "ArgumentsOptions",
+    "cli_arguments",
+]
