+++ src/robotframework_find_unused/reporter/base/argument_reporter.py
@@ -1,0 +1,35 @@
+from typing import TYPE_CHECKING
+
+from robotframework_find_unused.common.const import KeywordData
+
+from .partial.count_keywords import PartialReporter_CountKeywords
+from .partial.discover_files import PartialReporter_DiscoverFiles
+from .partial.keyword_definitions import (
+    PartialReporter_CustomKeywordDefinitions,
+    PartialReporter_DownloadedKeywordDefinitions,
+)
+from .partial.parse_files import PartialReporter_ParseFiles
+
+if TYPE_CHECKING:
+    from robotframework_find_unused.commands.arguments import ArgumentsOptions
+
+
+class ArgumentReporter(
+    PartialReporter_DiscoverFiles,
+    PartialReporter_ParseFiles,
+    PartialReporter_CustomKeywordDefinitions,
+    PartialReporter_DownloadedKeywordDefinitions,
+    PartialReporter_CountKeywords,
+):
+    """
+    Base reporter class for keyword command.
+    """
+
+    def __init__(self, options: "ArgumentsOptions") -> None:
+        self.options = options
+
+    def on_command_start(self):
+        """Before the command does anything"""
+
+    def on_command_end(self, counted_keywords: list[KeywordData]):
+        """When the command has done all the things"""
